chore: update sendxp
