--- src/commands/community/vote/vote_slash.ts
@@ -1,37 +1,0 @@
-import { SlashCommandBuilder } from "@discordjs/builders"
-import { SlashCommand } from "types/common"
-import { SLASH_PREFIX as PREFIX } from "utils/constants"
-import { composeEmbedMessage } from "utils/discordEmbed"
-import { disabledVoteEmbed } from "./"
-
-const command: SlashCommand = {
-  name: "vote",
-  category: "Community",
-  help: async () => {
-    return {
-      embeds: [
-        composeEmbedMessage(null, {
-          usage: `${PREFIX}vote`,
-          examples: `${PREFIX}vote\n${PREFIX}vote top\n${PREFIX}vote set #vote`,
-          includeCommandsList: true,
-        }),
-      ],
-    }
-  },
-  colorType: "Server",
-  prepare: () => {
-    return new SlashCommandBuilder()
-      .setDescription("Display voting streaks and links to vote")
-      .setName("vote")
-  },
-  ephemeral: true,
-  run: async () => {
-    return {
-      messageOptions: {
-        embeds: [disabledVoteEmbed()],
-      },
-    }
-  },
-}
-
-export default command
