+++ src/commands/sendxp/index/processor.ts
@@ -1,12 +1,9 @@
 import community from "adapters/community"
 import defi from "adapters/defi"
-import { CommandInteraction, Message } from "discord.js"
-import { APIError, GuildIdNotFoundError } from "errors"
-import { Command } from "types/common"
-import { getCommandArguments } from "utils/commands"
+import { Message, CommandInteraction } from "discord.js"
+import { APIError } from "errors"
+import { composeEmbedMessage, getErrorEmbed } from "ui/discord/embed"
 import { getEmoji, roundFloatNumber } from "utils/common"
-import { PREFIX } from "utils/constants"
-import { composeEmbedMessage, getErrorEmbed } from "utils/discordEmbed"
 
 export function getSendXPSuccessEmbed(
   recipientsId: string[],
@@ -81,60 +78,3 @@ export async function handleSendXp(
     },
   }
 }
-
-const command: Command = {
-  id: "sendxp",
-  command: "sendxp",
-  brief: "Send XP to members",
-  category: "Community",
-  run: async (msg) => {
-    if (!msg.guildId) {
-      throw new GuildIdNotFoundError({ message: msg })
-    }
-    // remove 'xp' argument
-    const args = getCommandArguments(msg).filter(
-      (arg) => arg.toLowerCase() !== "xp"
-    )
-    const each = args[args.length - 1].toLowerCase() === "each" ? true : false
-    const amount = Number(each ? args[args.length - 2] : args[args.length - 1])
-    if (Number.isNaN(amount)) {
-      return {
-        messageOptions: {
-          embeds: [
-            getErrorEmbed({
-              title: "Fail to send XP!",
-              description: "XP amount must be a number",
-            }),
-          ],
-        },
-      }
-    }
-
-    return handleSendXp(
-      msg,
-      args.slice(1, each ? -2 : -1).join(" "),
-      amount,
-      each
-    )
-  },
-  getHelpMessage: async (msg) => ({
-    embeds: [
-      composeEmbedMessage(msg, {
-        title: "Send XP to members",
-        usage: `${PREFIX}sendXP <recipient(s)> <amount> [each]`,
-        description: `You can send to recipients by:\n${getEmoji(
-          "POINTINGRIGHT"
-        )} Username(s): \`@tom\`, \`@john\`\n${getEmoji(
-          "POINTINGRIGHT"
-        )} Role(s): \`@dev\`, \`@staff\``,
-        examples: `${PREFIX}sendXP @john 5\n${PREFIX}sendXP @staff 5 XP`,
-      }),
-    ],
-  }),
-  colorType: "Server",
-  onlyAdministrator: true,
-  canRunWithoutAction: true,
-  minArguments: 3,
-}
-
-export default command
